Addressed more comments, removed oldreqs, renamed doi_builder, and moved DAMinimum model to models.py
